+++ src/routes/vpn/openvpn/openvpn.js
@@ -55,10 +55,8 @@ var router = express.Router();
  * @type ../../models/vpn/openvpn
  */
 
-import { PolicyRuleToOpenVPN } from '../../../models/policy/PolicyRuleToOpenVPN';
 import { Crt } from '../../../models/vpn/pki/Crt';
 import { OpenVPNPrefix } from '../../../models/vpn/openvpn/OpenVPNPrefix';
-import { PolicyCompilation } from '../../../models/policy/PolicyCompilation';
 import { OpenVPN } from '../../../models/vpn/openvpn/OpenVPN';
 import { Tree } from '../../../models/tree/Tree';
 const restrictedCheck = require('../../../middleware/restricted');
@@ -118,14 +116,6 @@ router.post('/', async(req, res) => {
 			await OpenVPNPrefix.applyOpenVPNPrefixes(req.dbCon,req.body.fwcloud,req.body.openvpn);
 		}
 
-		// Invalidate the compilation of the rules that use a prefix that use this new OpenVPN configuration.
-		let rules = await PolicyRuleToOpenVPN.searchOpenvpnInPrefixInRule(req.dbCon,req.body.fwcloud,cfg);
-		await PolicyCompilation.deleteRulesCompilation(req.body.fwcloud,rules);
-
-		// Invalidate the compilation of the rules that use a group that contains a prefix that use this new OpenVPN configuration.
-		let groups = await PolicyRuleToOpenVPN.searchOpenvpnInPrefixInGroup(req.dbCon,req.body.fwcloud,cfg);
-		await PolicyCompilation.deleteGroupsInRulesCompilation(req.dbCon,req.body.fwcloud,groups);
-
 		// If we are creaing an OpenVPN server configuration, then create the VPN virtual network interface with its assigned IP.
 		if (req.crt.type===2) // 1=Client certificate, 2=Server certificate.
 			await OpenVPN.createOpenvpnServerInterface(req,cfg);
@@ -143,18 +133,6 @@ router.post('/', async(req, res) => {
  */
 router.put('/', async(req, res) => {
 	try {
-		// Invalidate the compilation of the rules using this OpenVPN configuration.
-		let rules = await PolicyRuleToOpenVPN.searchOpenvpnInRule(req.dbCon,req.body.fwcloud,req.body.openvpn);
-		// Invalidate the compilation of the rules that use a prefix that use this OpenVPN configuration.
-		rules = rules.concat(await PolicyRuleToOpenVPN.searchOpenvpnInPrefixInRule(req.dbCon,req.body.fwcloud,req.body.openvpn));
-		await PolicyCompilation.deleteRulesCompilation(req.body.fwcloud,rules);
-		
-		// Invalidate the compilation of the rules that use a group that use this OpenVPN configuration.
-		let groups = await PolicyRuleToOpenVPN.searchOpenvpnInGroup(req.dbCon,req.body.fwcloud,req.body.openvpn);
-		// Invalidate the compilation of the rules that use a group that contains a prefix that use this OpenVPN configuration.
-		groups = groups.concat(await PolicyRuleToOpenVPN.searchOpenvpnInPrefixInGroup(req.dbCon,req.body.fwcloud,req.body.openvpn));
-		await PolicyCompilation.deleteGroupsInRulesCompilation(req.dbCon, req.body.fwcloud,groups);
-
 		await OpenVPN.updateCfg(req);
 
 		// First remove all the current configuration options.
@@ -275,14 +253,6 @@ router.put('/del',
 restrictedCheck.openvpn,
 async(req, res) => {
 	try {
-		// Invalidate the compilation of the rules that use a prefix that use this removed OpenVPN configuration.
-		let rules = await PolicyRuleToOpenVPN.searchOpenvpnInPrefixInRule(req.dbCon,req.body.fwcloud,req.body.openvpn);
-		await PolicyCompilation.deleteRulesCompilation(req.body.fwcloud,rules);
-
-		// Invalidate the compilation of the rules that use a group that contains a prefix that use this removed OpenVPN configuration.
-		let groups = await PolicyRuleToOpenVPN.searchOpenvpnInPrefixInGroup(req.dbCon,req.body.fwcloud,req.body.openvpn);
-		await PolicyCompilation.deleteGroupsInRulesCompilation(req.dbCon,req.body.fwcloud,groups);
-		
 		// Delete the configuration from de database.
 		await OpenVPN.delCfg(req.dbCon, req.body.fwcloud, req.body.openvpn);