+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bug in autodiscover when the `ip a` command returns interfaces with name like this one: `ens193.40@ens193:` In such cases the interface name must be the string preceding the `@` character.
 
 ### Changed
+- Removed table policy_c and all the code that uses it.
 - Removed `mysqldump` and `mysql-import` node modules.
 - Set the maximum size of accepted data for BodyParser to 2MB.
 - Disable confirmation token for `PUT /ping` API call.